int cast fix

+++ h_exec.c
@@ -76,11 +76,11 @@ int APIENTRY WinMain(HINSTANCE hInstance, HINSTANCE hPreviousInstance, LPSTR lpC
 				*(strrchr(szPath, '\\') + 1) = '\0';
 			strcat(szPath, szRun);
 
-			bRet = (int)ShellExecute(GetDesktopWindow(), NULL, szPath, szParam, NULL, SW_SHOWNORMAL);
+			bRet = ShellExecute(GetDesktopWindow(), NULL, szPath, szParam, NULL, SW_SHOWNORMAL);
 
 		} else {
 
-			bRet = (int)ShellExecute(GetDesktopWindow(), NULL, szRun, szParam, NULL, SW_SHOWNORMAL);
+			bRet = ShellExecute(GetDesktopWindow(), NULL, szRun, szParam, NULL, SW_SHOWNORMAL);
 		}
 
 	} else {
@@ -110,14 +110,14 @@ int APIENTRY WinMain(HINSTANCE hInstance, HINSTANCE hPreviousInstance, LPSTR lpC
 
 			strcat(szExec, szPgm);
 			if(ExistFile(szExec))
-				bRet = (int)ShellExecute(GetDesktopWindow(), "open", szExec, NULL, NULL, SW_SHOWNORMAL);
+				bRet = ShellExecute(GetDesktopWindow(), "open", szExec, NULL, NULL, SW_SHOWNORMAL);
 			else
-				bRet = (int)ShellExecute(GetDesktopWindow(), "open", szAlt, NULL, NULL, SW_SHOWNORMAL);
+				bRet = ShellExecute(GetDesktopWindow(), "open", szAlt, NULL, NULL, SW_SHOWNORMAL);
 		} else
 
 			// Must be a command line, then.
 
-			bRet = (int)ShellExecute(GetDesktopWindow(), "open", lpCmdLine, NULL, NULL, SW_SHOWNORMAL);
+			bRet = ShellExecute(GetDesktopWindow(), "open", lpCmdLine, NULL, NULL, SW_SHOWNORMAL);
 	}
 
 	if(bRet <= SE_ERROR) {
